fix package.json and update server.js
